--- .babelrc
@@ -1,11 +1,0 @@
-{
-	"presets": [
-		["@babel/preset-env", {
-			"targets": "last 2 versions"
-		}]
-	],
-	"plugins": [
-		"@babel/plugin-transform-runtime",
-		"@babel/plugin-proposal-class-properties"
-	]
-}
