+++ .gitignore
@@ -2,6 +2,9 @@
 node_modules/
 npm-debug.log
 lerna-debug.log
+package-lock.json
+yarn.lock
 
 # custom
 dist
+lib
