+++ packages/plugins/src/SearchBodyPlugin.ts
@@ -5,7 +5,7 @@ const TYPE_JSON = 'application/json'
 const TYPE_FORM = 'application/x-www-form-urlencoded'
 const TYPE_MULTIPART = 'multipart/form-data'
 
-export default (options) => async (ctx, next) => {
+export default (options?) => async (ctx, next) => {
 	const { request } = ctx
 	const { url, method, body, req: { search, type } } = request
 	if (search) {
@@ -26,7 +26,7 @@ export default (options) => async (ctx, next) => {
 			request.body = new URLSearchParams(body).toString()
 		} else if (request.is(TYPE_MULTIPART) && typeof FormData === 'function') {
 			const formData = new FormData()
-			Object.entries(body).forEach(([key, value]) => formData.append(key, value))
+			Object.entries(body).forEach(([key, value]: [string, string | Blob]) => formData.append(key, value))
 			request.body = formData
 		}
 	}
