+++ packages/web/.babelrc
@@ -1,0 +1,7 @@
+{
+	"presets": [
+		["@babel/preset-env", {
+			"targets": "last 2 versions"
+		}]
+	]
+}
