+++ packages/web/.npmignore
@@ -2,8 +2,11 @@
 node_modules
 .gitignore
 .babelrc
+tsconfig.json
 lerna-debug.log
+yarn.lock
 
 # custom
 test.js
+index.html
