+++ packages/web/index.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0, user-scalable=no">
+  <title>index</title>
+  <style type="text/css">
+    html, body { margin: 0; width: 100%; height: 100%; overflow: hidden; }
+    #body { width: 100%; height: 100%; display: -webkit-box; display: flex; }
+  </style>
+</head>
+<body>
+  <div id="body"></div>
+  <script src="dist/index.js"></script>
+  <script type="text/javascript">
+    const fetch = FetchPlus.fetch
+
+    fetch.use(async (ctx, next) => {
+    const { request } = ctx
+      request.url = `https://unpkg.com/${request.url}`
+      await next()
+    })
+
+    fetch('react', { 
+      // dataType: 'json',
+      search: { meta: true }
+    }).then(data => {
+      console.log(data)
+    })
+  </script>
+</body>
+</html>
